+++ lib/src/internal/entity_builder.dart
@@ -1,10 +1,13 @@
+
 import 'package:restrr/restrr.dart';
 import 'package:restrr/src/internal/restrr_impl.dart';
 
+import 'entities/account_impl.dart';
 import 'entities/currency/currency_impl.dart';
 import 'entities/currency/custom_currency_impl.dart';
 import 'entities/session/partial_session_impl.dart';
 import 'entities/session/session_impl.dart';
+import 'entities/transaction_impl.dart';
 import 'entities/user_impl.dart';
 
 /// Defines how to build entities from JSON responses.
@@ -59,6 +62,37 @@ class EntityBuilder {
       );
     }
     return session;
+  }
+
+  Account buildAccount(Map<String, dynamic> json) {
+    final AccountImpl account = AccountImpl(
+      api: api,
+      id: json['id'],
+      name: json['name'],
+      description: json['description'],
+      iban: json['iban'],
+      balance: json['balance'],
+      originalBalance: json['original_balance'],
+      currencyId: json['currency_id'],
+      createdAt: DateTime.parse(json['created_at']),
+    );
+    return api.accountCache.cache(account);
+  }
+
+  Transaction buildTransaction(Map<String, dynamic> json) {
+    final TransactionImpl transaction = TransactionImpl(
+      api: api,
+      id: json['id'],
+      sourceId: json['source_id'],
+      destinationId: json['destination_id'],
+      amount: json['amount'],
+      currencyId: json['currency'],
+      description: json['description'],
+      budgetId: json['budget_id'],
+      createdAt: DateTime.parse(json['created_at']),
+      executedAt: DateTime.parse(json['executed_at']),
+    );
+    return api.transactionCache.cache(transaction);
   }
 
   User buildUser(Map<String, dynamic> json) {