+++ lib/src/internal/restrr_impl.dart
@@ -20,10 +20,14 @@ class RestrrImpl implements Restrr {
 
   late final EntityCacheView<Currency> currencyCache = EntityCacheView(this);
   late final EntityCacheView<PartialSession> sessionCache = EntityCacheView(this);
+  late final EntityCacheView<Account> accountCache = EntityCacheView(this);
+  late final EntityCacheView<Transaction> transactionCache = EntityCacheView(this);
   late final EntityCacheView<User> userCache = EntityCacheView(this);
 
   late final PageCacheView<Currency> currencyPageCache = PageCacheView(this);
   late final PageCacheView<PartialSession> sessionPageCache = PageCacheView(this);
+  late final PageCacheView<Account> accountPageCache = PageCacheView(this);
+  late final PageCacheView<Transaction> transactionPageCache = PageCacheView(this);
 
   RestrrImpl({required this.routeOptions, required Map<Type, Function> eventMap, this.options = const RestrrOptions()})
       : eventHandler = RestrrEventHandler(eventMap);
@@ -100,6 +104,53 @@ class RestrrImpl implements Restrr {
     return response.hasData && response.data!;
   }
 
+  /* Accounts */
+
+  @override
+  Future<Account> createAccount(
+      {required String name,
+      required int originalBalance,
+      required Id currency,
+      String? description,
+      String? iban}) async {
+    final RestResponse<Account> response = await requestHandler
+        .apiRequest(route: AccountRoutes.create.compile(), mapper: (json) => entityBuilder.buildAccount(json), body: {
+      'name': name,
+      'original_balance': originalBalance,
+      'currency_id': currency,
+      if (description != null) 'description': description,
+      if (iban != null) 'iban': iban
+    });
+    if (response.hasError) {
+      throw response.error!;
+    }
+    return accountCache.cache(response.data!);
+  }
+
+  @override
+  List<Account> getAccounts() => accountCache.getAll();
+
+  @override
+  Future<Account> retrieveAccountById(Id id, {bool forceRetrieve = false}) async {
+    return RequestUtils.getOrRetrieveSingle(
+        key: id,
+        cacheView: accountCache,
+        compiledRoute: AccountRoutes.getById.compile(params: [id]),
+        mapper: (json) => entityBuilder.buildAccount(json),
+        forceRetrieve: forceRetrieve);
+  }
+
+  @override
+  Future<Paginated<Account>> retrieveAllAccounts({int page = 1, int limit = 25, bool forceRetrieve = false}) async {
+    return RequestUtils.getOrRetrievePage(
+        pageCache: accountPageCache,
+        compiledRoute: AccountRoutes.getAll.compile(),
+        page: page,
+        limit: limit,
+        mapper: (json) => entityBuilder.buildAccount(json),
+        forceRetrieve: forceRetrieve);
+  }
+
   /* Currencies */
 
   @override
@@ -118,9 +169,11 @@ class RestrrImpl implements Restrr {
       throw response.error!;
     }
     // invalidate cache
-    currencyCache.clear();
-    return response.data!;
+    return currencyCache.cache(response.data!);
   }
+
+  @override
+  List<Currency> getCurrencies() => currencyCache.getAll();
 
   @override
   Future<Currency> retrieveCurrencyById(Id id, {bool forceRetrieve = false}) async {
@@ -140,6 +193,62 @@ class RestrrImpl implements Restrr {
         page: page,
         limit: limit,
         mapper: (json) => entityBuilder.buildCurrency(json),
+        forceRetrieve: forceRetrieve);
+  }
+
+  /* Transactions */
+
+  @override
+  Future<Transaction> createTransaction(
+      {required int amount,
+      required Id currencyId,
+      required DateTime executedAt,
+      String? description,
+      Id? sourceId,
+      Id? destinationId,
+      Id? budgetId}) async {
+    if (sourceId == null && destinationId == null) {
+      throw ArgumentError('Either source or destination must be set!');
+    }
+    final RestResponse<Transaction> response = await requestHandler.apiRequest(
+        route: TransactionRoutes.create.compile(),
+        mapper: (json) => entityBuilder.buildTransaction(json),
+        body: {
+          'amount': amount,
+          'currency_id': currencyId,
+          'executed_at': executedAt.toUtc().toIso8601String(),
+          if (description != null) 'description': description,
+          if (sourceId != null) 'source_id': sourceId,
+          if (destinationId != null) 'destination_id': destinationId,
+          if (budgetId != null) 'budget_id': budgetId
+        });
+    if (response.hasError) {
+      throw response.error!;
+    }
+    // invalidate cache
+    transactionCache.clear();
+    return response.data!;
+  }
+
+  @override
+  Future<Transaction> retrieveTransactionById(Id id, {bool forceRetrieve = false}) async {
+    return RequestUtils.getOrRetrieveSingle(
+        key: id,
+        cacheView: transactionCache,
+        compiledRoute: TransactionRoutes.getById.compile(params: [id]),
+        mapper: (json) => entityBuilder.buildTransaction(json),
+        forceRetrieve: forceRetrieve);
+  }
+
+  @override
+  Future<Paginated<Transaction>> retrieveAllTransactions(
+      {int page = 1, int limit = 25, bool forceRetrieve = false}) async {
+    return RequestUtils.getOrRetrievePage(
+        pageCache: transactionPageCache,
+        compiledRoute: TransactionRoutes.getAll.compile(),
+        page: page,
+        limit: limit,
+        mapper: (json) => entityBuilder.buildTransaction(json),
         forceRetrieve: forceRetrieve);
   }
 }